chagne pkg name after rename directory
