+++ api/go/src/lmm/api/storage/file/local_asset_uploader.go
@@ -1,4 +1,4 @@
-package uploader
+package file
 
 import (
 	"bufio"
